run inetgration tests as junit test suite
