+++ src/app/build.gradle
@@ -106,5 +106,8 @@ task copySpec(type: Copy) {
 
 test.dependsOn 'copySpec'
 
+test.exclude "**/integration_tests/tests/**"
+
+
 copyConfig.finalizedBy 'showConfig'
 copyTestConfig.finalizedBy 'showTestConfig'
